// src/core/config/env-variables.enum.ts
export enum NodeEnv {
  DEVELOPMENT = 'development',
  STAGING = 'staging',
  PRODUCTION = 'production',
  TEST = 'test',
}

export enum EnvVariableName {
  NODE_ENV = 'NODE_ENV',
  PORT = 'PORT',
  MONGO_DB_URL = 'MONGO_DB_URL',
  JWT_ACCESS_TOKEN_SECRET = 'JWT_ACCESS_TOKEN_SECRET',
  JWT_REFRESH_TOKEN_SECRET = 'JWT_REFRESH_TOKEN_SECRET',
  JWT_ACCESS_TOKEN_TTL = 'JWT_ACCESS_TOKEN_TTL',
  JWT_REFRESH_TOKEN_TTL = 'JWT_REFRESH_TOKEN_TTL',
  APP_EMAIL_ADDRESS = 'APP_EMAIL_ADDRESS',
  APP_EMAIL_PASSWORD = 'APP_EMAIL_PASSWORD',
  INCLUDE_TESTING_MODULE = 'INCLUDE_TESTING_MODULE',
  POSTGRES_DB = 'POSTGRES_DB',
  POSTGRES_USER = 'POSTGRES_USER',
  POSTGRES_PASSWORD = 'POSTGRES_PASSWORD',
  POSTGRES_PORT = 'POSTGRES_PORT',
}